refactor(core): abstract cache and context [no changelog]
