+++ core/src/all_modules.py
@@ -47,6 +47,10 @@ storage
 import storage
 storage.cache
 import storage.cache
+storage.cache_codec
+import storage.cache_codec
+storage.cache_common
+import storage.cache_common
 storage.common
 import storage.common
 storage.debug
@@ -201,12 +205,20 @@ trezor.utils
 import trezor.utils
 trezor.wire
 import trezor.wire
-trezor.wire.codec_v1
-import trezor.wire.codec_v1
+trezor.wire.codec
+import trezor.wire.codec
+trezor.wire.codec.codec_context
+import trezor.wire.codec.codec_context
+trezor.wire.codec.codec_v1
+import trezor.wire.codec.codec_v1
 trezor.wire.context
 import trezor.wire.context
 trezor.wire.errors
 import trezor.wire.errors
+trezor.wire.message_handler
+import trezor.wire.message_handler
+trezor.wire.protocol_common
+import trezor.wire.protocol_common
 trezor.workflow
 import trezor.workflow
 apps
@@ -309,6 +321,8 @@ apps.common.backup
 import apps.common.backup
 apps.common.backup_types
 import apps.common.backup_types
+apps.common.cache
+import apps.common.cache
 apps.common.cbor
 import apps.common.cbor
 apps.common.coininfo
