+++ core/src/trezor/wire/codec/codec_v1.py
@@ -3,6 +3,7 @@ from micropython import const
 from typing import TYPE_CHECKING
 
 from trezor import io, loop, utils
+from trezor.wire.protocol_common import Message, WireError
 
 if TYPE_CHECKING:
     from trezorio import WireInterface
@@ -16,14 +17,8 @@ _REP_INIT_DATA = const(9)  # offset of data in the initial report
 _REP_CONT_DATA = const(1)  # offset of data in the continuation report
 
 
-class CodecError(Exception):
+class CodecError(WireError):
     pass
-
-
-class Message:
-    def __init__(self, mtype: int, mdata: bytes) -> None:
-        self.type = mtype
-        self.data = mdata
 
 
 async def read_message(iface: WireInterface, buffer: utils.BufferType) -> Message:
